--- cloud_config/build.gradle
@@ -1,43 +1,0 @@
-plugins {
-	id 'java'
-	id 'org.springframework.boot' version '3.0.0'
-	id 'io.spring.dependency-management' version '1.1.0'
-}
-
-group = 'ru.clevertec'
-version = '0.0.1-SNAPSHOT'
-sourceCompatibility = '17'
-
-repositories {
-	mavenCentral()
-	maven { url 'https://repo.spring.io/milestone' }
-}
-
-ext {
-	set('springCloudVersion', "2022.0.0-RC2")
-}
-
-dependencies {
-	implementation 'org.springframework.boot:spring-boot-starter-data-jpa'
-	implementation 'org.springframework.cloud:spring-cloud-starter-config'
-	testImplementation 'org.springframework.boot:spring-boot-starter-test'
-	implementation group: 'org.postgresql', name: 'postgresql', version: '42.1.4'
-
-	implementation group: 'org.springframework.boot', name: 'spring-boot-autoconfigure', version: '3.0.0'
-
-	implementation group: 'org.springframework.cloud', name: 'spring-cloud-config-server', version: '4.0.0'
-
-	implementation group: 'org.springframework.boot', name: 'spring-boot-starter-data-jpa', version: '3.0.0'
-
-}
-
-dependencyManagement {
-	imports {
-		mavenBom "org.springframework.cloud:spring-cloud-dependencies:${springCloudVersion}"
-
-	}
-}
-
-tasks.named('test') {
-	useJUnitPlatform()
-}
