--- cloud_config/settings.gradle
@@ -1,1 +1,0 @@
-rootProject.name = 'cloud_config'
