--- cloud_config/src/main/java/ru/clevertec/cloud_config/CloudConfigApplication.java
@@ -1,15 +1,0 @@
-package ru.clevertec.cloud_config;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.cloud.config.server.EnableConfigServer;
-
-@EnableConfigServer
-@SpringBootApplication
-public class CloudConfigApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(CloudConfigApplication.class, args);
-	}
-
-}
