--- cloud_config/src/main/resources/application.yml
@@ -1,10 +1,0 @@
-spring:
-  profiles:
-    include: native
-  application:
-    name: setting-api
-  cloud:
-    config:
-      server:
-        native:
-          search-locations: classpath:configs/,classpath:configs/{application}
