--- cloud_config/src/main/resources/configs/user-api/user-api.yml
@@ -1,14 +1,0 @@
-spring:
-  application:
-    name: user-api
-  jpa:
-    hibernate:
-      ddl-auto: create
-    database-platform: org.hibernate.dialect.PostgreSQLDialect
-    show-sql: true
-  datasource:
-    url: jdbc:postgresql://localhost:5432/artemaverkov
-    username: root
-    password: root
-    driver-class-name: org.postgresql.Driver
-
