--- cloud_config/src/test/java/ru/clevertec/cloud_config/CloudConfigApplicationTests.java
@@ -1,13 +1,0 @@
-package ru.clevertec.cloud_config;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class CloudConfigApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
