+++ testWork/build.gradle
@@ -2,11 +2,20 @@ plugins {
 	id 'java'
 	id 'org.springframework.boot' version '3.0.0'
 	id 'io.spring.dependency-management' version '1.1.0'
+	id "com.sidneysimmons.gradle-plugin-docker" version "0.0.7"
+	id "vast.gradle-docker" version "1.3"
+
 }
+
+apply plugin: 'application'
+mainClassName = 'ru.clevertec.testWork.TestWorkApplication'
 
 group = 'ru.clevertec'
 version = '0.0.1-SNAPSHOT'
 sourceCompatibility = '17'
+
+
+
 
 configurations {
 	compileOnly {
@@ -15,7 +24,9 @@ configurations {
 }
 
 repositories {
+	mavenLocal()
 	mavenCentral()
+
 }
 
 dependencies {
@@ -51,10 +62,14 @@ dependencies {
 	testImplementation group: 'org.mockito', name: 'mockito-core', version: '4.10.0'
 // https://mvnrepository.com/artifact/org.junit.platform/junit-platform-runner
 	testImplementation group: 'org.junit.platform', name: 'junit-platform-runner', version: '1.9.1'
+	}
 
 
-}
 
 tasks.named('test') {
 	useJUnitPlatform()
+}
+jar {
+	baseName = 'gs-gradle'
+	version =  '0.1.0'
 }
