+++ testWork/docker/services/app/Dockerfile
@@ -1,0 +1,4 @@
+FROM openjdk:17-jdk
+WORKDIR /usr/local/app
+ADD testWork-0.0.1-SNAPSHOT.jar app.jar
+CMD java -jar ./app.jar
