+++ testWork/docker/services/postgres/Dockerfile
@@ -1,0 +1,1 @@
+FROM postgres:12-alpine
