+++ testWork/src/main/java/ru/clevertec/testWork/entities/discount/Discount.java
@@ -1,6 +1,8 @@
 package ru.clevertec.testWork.entities.discount;
 
 import jakarta.persistence.Entity;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
@@ -12,11 +14,12 @@ import org.hibernate.annotations.OptimisticLocking;
 @Data
 @Builder
 @Entity
-@OptimisticLocking(type = OptimisticLockType.VERSION)
 @AllArgsConstructor
 @NoArgsConstructor
 public class Discount {
     @Id
-    private long id;
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
     private String name;
+
 }
