+++ testWork/src/main/java/ru/clevertec/testWork/service/discount/DiscountApiService.java
@@ -4,9 +4,12 @@ import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 import ru.clevertec.testWork.dto.discount.DiscountDto;
 import ru.clevertec.testWork.entities.discount.Discount;
+
 import ru.clevertec.testWork.repository.discount.DiscountRepository;
 
 import java.util.List;
+import java.util.NoSuchElementException;
+import java.util.Objects;
 
 
 @Service
@@ -21,16 +24,28 @@ public record DiscountApiService
 
     @Override
     public Discount read(long id) {
-        return discountRepository.findById(id).get();
+        return discountRepository.findById(id).orElseThrow(NoSuchElementException::new);
     }
 
     @Override
     public boolean update(DiscountDto discountDto, Long id) {
+        Discount read = read(id);
+        if (Objects.nonNull(read)) {
+            Discount discount = buildDiscount(discountDto);
+            discount.setId(id);
+            discountRepository.save(discount);
+            return true;
+        }
         return false;
     }
 
     @Override
     public boolean delete(Long id) {
+        Discount read = read(id);
+        if (Objects.nonNull(read)) {
+            discountRepository.deleteById(id);
+            return true;
+        }
         return false;
     }
