+++ testWork/src/main/resources/application.properties
@@ -1,9 +1,9 @@
-spring.datasource.url=jdbc:postgresql://localhost:5432/artemaverkov
+spring.datasource.url=jdbc:postgresql://localhost:5432/root
 spring.datasource.username=artemaverkov
 spring.datasource.password=root
 spring.datasource.driver-class-name=org.postgresql.Driver
 spring.jpa.database=postgresql
 spring.jpa.database-platform=org.hibernate.dialect.PostgreSQLDialect
-spring.jpa.hibernate.ddl-auto=update
+spring.jpa.hibernate.ddl-auto=create
 spring.jpa.show-sql = true
-server.port = 8899
+server.port = 8080
