+++ testWork/src/test/java/ru/clevertec/testWork/TestWorkApplicationTests.java
@@ -1,13 +1,9 @@
 package ru.clevertec.testWork;
 
-import org.junit.jupiter.api.Test;
+
 import org.springframework.boot.test.context.SpringBootTest;
 
 @SpringBootTest
 class TestWorkApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
 
 }
