+++ testWork/src/test/java/ru/clevertec/testWork/common/extension/discount/ValidParameterResolverDiscount.java
@@ -1,5 +1,4 @@
 package ru.clevertec.testWork.common.extension.discount;
-
 import org.junit.jupiter.api.extension.ExtensionContext;
 import org.junit.jupiter.api.extension.ParameterContext;
 import org.junit.jupiter.api.extension.ParameterResolutionException;
