+++ testWork/src/test/java/ru/clevertec/testWork/common/service/product/ProductServiceImplTest.java
@@ -1,6 +1,5 @@
 package ru.clevertec.testWork.common.service.product;
 
-
 import jakarta.persistence.EntityNotFoundException;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Nested;
